+++ tests/kernel/arm_runtime_nmi/prj.conf
@@ -1,1 +1,2 @@
 CONFIG_RUNTIME_NMI=y
+CONFIG_ZTEST=y
