+++ tests/kernel/arm_runtime_nmi/src/Makefile
@@ -1,3 +1,4 @@
 ccflags-y += -I${ZEPHYR_BASE}/tests/include
+include ${ZEPHYR_BASE}/tests/Makefile.test
 
-obj-y = main.o
+obj-y = main.o arm_runtime_nmi.o
